fixing linting errors
